requirements.txt
google-generativeai>=0.3.0
sentence-transformers
tqdm
numpy
python-dotenv
torch
flask
flask-cors
